readme: batch-scan merged PR reward intake
