+++ .pm/tasks/task_b21f68180cf64edb9ab5dc50bb896ed5.execution.md
@@ -1,0 +1,17 @@
+# task_b21f68180cf64edb9ab5dc50bb896ed5 Execution Log
+
+- task_uid: task_b21f68180cf64edb9ab5dc50bb896ed5
+- title: TASK-README-070 为已合入 PR 奖励 intake 增加按轮扫描脚本
+- owner_role: liveops_community
+- worktree_hint: oasis7-readme-reward-merged-pr-round-scan
+
+<!-- Append entries using:
+Example:
+  ## YYYY-MM-DD HH:MM:SS CST / role_name
+  - 完成内容: ...
+  - 遗留事项: ...
+-->
+## 2026-04-13 10:30:12 CST / liveops_community
+- 完成内容: 为 `TASK-README-070` 新增 merged PR reward round scan 闭环，补齐共享解析库 `scripts/readme_reward_pr_intake_lib.py`、批量扫描脚本 `scripts/readme-reward-pr-intake-round-scan.py` 与 smoke `scripts/readme-reward-pr-intake-round-scan-smoke.sh`；同时让 `scripts/readme-reward-pr-intake-import.py` 复用共享 contract，并将 `doc/readme/prd.md`、`doc/readme/project.md` 与 reward pack / ledger 专题同步回写到“按 merged 时间窗扫描 -> 输出 `ready/deferred/no_reward_review_requested/invalid_intake` -> 再导入 round ledger”的正式流程。另补齐 `.pm/roles/viewer_engineer/backlog/committed.yaml` 缺失的标准头，解除现有 `pm-lint` 基线阻断。
+- 验证结果: 已执行 `python3 -m py_compile scripts/readme_reward_pr_intake_lib.py scripts/readme-reward-pr-intake-import.py scripts/readme-reward-pr-intake-round-scan.py`、`./scripts/readme-reward-pr-intake-import-smoke.sh`、`./scripts/readme-reward-pr-intake-round-scan-smoke.sh`、`rg -n "PRD-README-044|Flow-RM-023|AC-31|DEC-RM-042|TASK-README-070|merged PR reward round scan" doc/readme/prd.md doc/readme/project.md doc/readme/governance/readme-limited-preview-contributor-reward-ledger-2026-03-22.prd.md doc/readme/governance/readme-limited-preview-contributor-reward-ledger-2026-03-22.project.md doc/readme/governance/readme-limited-preview-contributor-reward-ledger-2026-03-22.md`、`python3 scripts/pm/pm_store.py task-execution-log-lint .`、`./scripts/pm/lint.sh`、`./scripts/doc-governance-check.sh` 与 `git diff --check`，结果通过。
+- 遗留事项: 当前仅完成 merged PR 的批量 intake 归集与 ledger 候选生成；真实发放仍需后续 round ledger 填写、`producer_system_designer` 审批，以及 reward reserve 地址绑定和 QA `pass` 等既有治理门禁闭环。
